% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffRelRisk.R
\name{DRRCI}
\alias{DRRCI}
\title{Difference in Relative Risks}
\usage{
DRRCI(x1 = NA, x2 = NA, x3 = NA, x4 = NA, n1 = NA, n2 = NA,
  n3 = NA, n4 = NA, tr = NA, Y = NA, Z = NA, alpha = 0.05,
  altParam = 0, deltaMethod = 0, refPop = NA, LRT = 0, estRR = 0)
}
\arguments{
\item{x1}{The number of events in the first group  (either a single number if there is one strata or a vector if there are multiple strata)}

\item{x2}{The number of events in the second group (either a single number if there is one strata or a vector if there are multiple strata)}

\item{x3}{The number of events in the third group (either a single number if there is one strata or a vector if there are multiple strata)}

\item{x4}{The number of events in the fourth group if the study included four groups (either missing if a three-group study, a single number if there is one strata or a vector if there are multiple strata; note: not all options are currently available when there are four groups)}

\item{n1}{The total number of subjects in the first group  (either a single number if there is one strata or a vector if there are multiple strata)}

\item{n2}{The total number of subjects in the second group (either a single number if there is one strata or a vector if there are multiple strata)}

\item{n3}{The total number of subjects in the third group (either a single number if there is one strata or a vector if there are multiple strata)}

\item{n4}{The total number of subjects in the fourth group if the study included four groups (either missing if a three-group study, a single number if there is one strata or a vector if there are multiple strata; note: not all options are currently available when there are four groups)}

\item{tr}{The treatment group for each subject (a vector with one value per subject; values are either 0, 1, or 2; only required if want to adjust for covariates)}

\item{Y}{The outcome for each subject (a vector with one value per subject; values are either 0 or 1; only required if want to adjust for covariates)}

\item{Z}{The covariates for each subject (a matrix with one value per subject and one column per covariate; only required if want to adjust for covariates)}

\item{alpha}{The two-sided error rate for the confidence interval (i.e. 0.05 for a 95\% confidence interval)}

\item{altParam}{A binary variable indicating whether to use the full likelihood (altParam=0) or assume the log-linear model holds true (altParam==1) (Note: altParam=1 is equivalent to entering subject Y, tr, Z w/ Z set to binary indicators for strata)}

\item{deltaMethod}{A binary variable indicating whether the delta method should be used (note: if deltaMethod==1, then default refPop="All")}

\item{refPop}{The target population when standardization is used (options are "All" -- all subjects, "Treat" -- treated population, "Cont" -- control population; "Opt" -- an optimally chosen population; "Ind" -- indirect standardization for the delta method)}

\item{LRT}{A binary variable indicating whether a likelihood-based confidence interval should be used (set to 0 for score-based confidence interval; set to 1 for likelihood-based confidence interval)}

\item{estRR}{A binary variable indicating whether to calculate the confidence intervals for each relative risk, as opposed to the difference (set to 0 for calculating the CI for difference, 1 for calculating the CIs for RR; note only works when data is entered as Z, Y, tr)}
}
\value{
EST The estimated difference in relative risks

LB The lower bound for the difference

UB The upper bound for the difference

beta The MLE (and asymptotic SE) for the coefficients for individual level covariates (when adjusting for covariates)
}
\description{
This function estimates and provides a confidence interval for the difference between
two relative risks. The first option is there are three populations and the goal to estimate
the difference: p2/p3 - p1/p3, where p1, p2, and p3 are the risks in each population.
The second option is there are four populations and the goal is to estimate the difference:
p2/p4 - p1/p3.
}
\examples{
DRRCI(x1=20,x2=50,x3=100,n1=1000,n2=1000,n3=1000)
#The simplest case where I have three groups and
#want to estimate (50/1000 - 20/1000)/(100/1000)
##
DRRCI(x1=c(15,5),x2=c(30,20),x3=c(50,50),n1=c(600,400),n2=c(400,600),n3=c(500,500))
#I now have observed individuals in two strata
#here I could use altParam = 1 if I wanted to assume that log-linear model held
#here I could use refPop   = "All" if I wanted to use direct standardization
##
DRRCI(x1=20,x2=50,x3=100,x4=90,n1=1000,n2=1000,n3=1000,n4=90)
#I now have a separate "control" group for each treatment group
##
tr = rep(c(0,1,2),each=100)
Y  = c(rep(1,30),rep(0,70),rep(1,10),rep(0,90),rep(1,20),rep(0,80))
Z  = cbind(1,rnorm(300),rnorm(300))
DRRCI(tr=tr,Y=Y,Z=Z)
#I now adjust for covariates
}
